// ui/src/lib/api/generated/model/project.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire API
 * AI-powered fashion platform API
 * OpenAPI spec version: 1.0.0
 */

export interface Project {
  _id: string;
  userId: string;
  name: string;
  description: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}